factor synthesis stuff to another file
