+++ ngapp/src/lib/synth.ts
@@ -1,0 +1,9 @@
+import PlaybackHandle from "./synth/playbackHandle";
+import showButtons from "./synth/showButtons";
+import Buttons from "./synth/buttons";
+
+const synth = {
+  PlaybackHandle, showButtons, Buttons,
+}
+
+export default synth;
